scripts/requirements.txt
requests
google-genai
supabase
python-dotenv
numpy
